+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/validation/core/DccValidator.kt
@@ -3,8 +3,8 @@ package de.rki.coronawarnapp.covidcertificate.validation.core
 import de.rki.coronawarnapp.covidcertificate.common.certificate.DccData
 import de.rki.coronawarnapp.covidcertificate.common.certificate.DccJsonSchemaValidator
 import de.rki.coronawarnapp.covidcertificate.common.certificate.DccV1
+import de.rki.coronawarnapp.covidcertificate.validation.core.business.BusinessValidator
 import de.rki.coronawarnapp.covidcertificate.validation.core.country.DccCountry
-import de.rki.coronawarnapp.covidcertificate.validation.core.validation.business.BusinessValidator
 import org.joda.time.Instant
 import timber.log.Timber
 import javax.inject.Inject
@@ -23,8 +23,6 @@ class DccValidator @Inject constructor(
         certificate: DccData<out DccV1.MetaData>,
     ): DccValidation {
         Timber.tag(TAG).v("validateDcc(country=%s)", arrivalCountry)
-
-        // TODO update value sets, acceptance rules, and invalidation rules
 
         val expirationCheckPassed = certificate.expiresAfter(arrivalTime)
         val jsonSchemaCheckPassed = dccJsonSchemaValidator.isValid(certificate.certificateJson).isValid
