+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/validation/core/business/BusinessValidation.kt
@@ -1,0 +1,8 @@
+package de.rki.coronawarnapp.covidcertificate.validation.core.business
+
+import de.rki.coronawarnapp.covidcertificate.validation.core.rule.EvaluatedDccRule
+
+data class BusinessValidation(
+    val acceptanceRules: Set<EvaluatedDccRule>,
+    val invalidationRules: Set<EvaluatedDccRule>
+)
