+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/validation/core/business/BusinessValidator.kt
@@ -1,14 +1,14 @@
-package de.rki.coronawarnapp.covidcertificate.validation.core.validation.business
+package de.rki.coronawarnapp.covidcertificate.validation.core.business
 
 import dagger.Reusable
 import de.rki.coronawarnapp.covidcertificate.common.certificate.DccData
+import de.rki.coronawarnapp.covidcertificate.common.certificate.DccJsonSchema
 import de.rki.coronawarnapp.covidcertificate.common.certificate.DccV1
 import de.rki.coronawarnapp.covidcertificate.validation.core.DccValidationRepository
+import de.rki.coronawarnapp.covidcertificate.validation.core.business.wrapper.CertLogicEngineWrapper
+import de.rki.coronawarnapp.covidcertificate.validation.core.business.wrapper.filterRelevantRules
+import de.rki.coronawarnapp.covidcertificate.validation.core.business.wrapper.type
 import de.rki.coronawarnapp.covidcertificate.validation.core.country.DccCountry
-import de.rki.coronawarnapp.covidcertificate.validation.core.validation.EvaluatedDccRule
-import de.rki.coronawarnapp.covidcertificate.validation.core.validation.wrapper.CertLogicEngineWrapper
-import de.rki.coronawarnapp.covidcertificate.validation.core.validation.wrapper.filterRelevantRules
-import de.rki.coronawarnapp.covidcertificate.validation.core.validation.wrapper.type
 import org.joda.time.Instant
 import javax.inject.Inject
 
@@ -16,12 +16,17 @@ import javax.inject.Inject
 class BusinessValidator @Inject constructor(
     private val certLogicEngineWrapper: CertLogicEngineWrapper,
     private val ruleRepository: DccValidationRepository,
+    private val dccJsonSchema: DccJsonSchema,
 ) {
     suspend fun validate(
         arrivalCountry: DccCountry,
         validationClock: Instant,
         certificate: DccData<out DccV1.MetaData>,
     ): BusinessValidation {
+
+        // TODO update value sets
+
+        ruleRepository.refresh()
 
         // accepted by arrival country
         val acceptanceResults = certLogicEngineWrapper.process(
@@ -31,7 +36,8 @@ class BusinessValidator @Inject constructor(
             ),
             validationClock = validationClock,
             certificate = certificate,
-            countryCode = arrivalCountry.countryCode
+            countryCode = arrivalCountry.countryCode,
+            schemaJson = dccJsonSchema.rawSchema
         )
 
         // valid as defined by the issuing country
@@ -43,12 +49,13 @@ class BusinessValidator @Inject constructor(
             ),
             validationClock = validationClock,
             certificate = certificate,
-            countryCode = issuerCountry.countryCode
+            countryCode = issuerCountry.countryCode,
+            schemaJson = dccJsonSchema.rawSchema
         )
 
-        return object : BusinessValidation {
-            override val acceptanceRules: Set<EvaluatedDccRule> = acceptanceResults
-            override val invalidationRules: Set<EvaluatedDccRule> = invalidationResults
-        }
+        return BusinessValidation(
+            acceptanceRules = acceptanceResults,
+            invalidationRules = invalidationResults
+        )
     }
 }
