+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/validation/core/business/wrapper/CertLogicEngineWrapper.kt
@@ -1,20 +1,17 @@
-package de.rki.coronawarnapp.covidcertificate.validation.core.validation.wrapper
+package de.rki.coronawarnapp.covidcertificate.validation.core.business.wrapper
 
 import dagger.Reusable
 import de.rki.coronawarnapp.covidcertificate.common.certificate.DccData
-import de.rki.coronawarnapp.covidcertificate.common.certificate.DccJsonSchema
 import de.rki.coronawarnapp.covidcertificate.common.certificate.DccV1
 import de.rki.coronawarnapp.covidcertificate.validation.core.rule.DccValidationRule
-import de.rki.coronawarnapp.covidcertificate.validation.core.validation.EvaluatedDccRule
+import de.rki.coronawarnapp.covidcertificate.validation.core.rule.EvaluatedDccRule
 import dgca.verifier.app.engine.DefaultCertLogicEngine
 import dgca.verifier.app.engine.DefaultJsonLogicValidator
 import org.joda.time.Instant
 import javax.inject.Inject
 
 @Reusable
-class CertLogicEngineWrapper @Inject constructor(
-    private val dccJsonSchema: DccJsonSchema,
-) {
+class CertLogicEngineWrapper @Inject constructor() {
 
     private val engine: DefaultCertLogicEngine by lazy {
         DefaultCertLogicEngine(DefaultJsonLogicValidator())
@@ -25,9 +22,10 @@ class CertLogicEngineWrapper @Inject constructor(
         validationClock: Instant,
         certificate: DccData<out DccV1.MetaData>,
         countryCode: String,
+        schemaJson: String
     ): Set<EvaluatedDccRule> = engine.validate(
         hcertVersionString = certificate.certificate.version,
-        schemaJson = dccJsonSchema.rawSchema,
+        schemaJson = schemaJson,
         rules = rules.map { it.asExternalRule },
         externalParameter = assembleExternalParameter(certificate, validationClock, countryCode),
         payload = certificate.certificateJson
