+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/validation/core/rule/EvaluatedDccRule.kt
@@ -1,0 +1,6 @@
+package de.rki.coronawarnapp.covidcertificate.validation.core.rule
+
+data class EvaluatedDccRule(
+    val rule: DccValidationRule,
+    val result: DccValidationRule.Result,
+)
