--- Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/validation/core/validation/EvaluatedDccRule.kt
@@ -1,8 +1,0 @@
-package de.rki.coronawarnapp.covidcertificate.validation.core.validation
-
-import de.rki.coronawarnapp.covidcertificate.validation.core.rule.DccValidationRule
-
-data class EvaluatedDccRule(
-    val rule: DccValidationRule,
-    val result: DccValidationRule.Result,
-)
