--- Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/validation/core/validation/business/BusinessValidation.kt
@@ -1,8 +1,0 @@
-package de.rki.coronawarnapp.covidcertificate.validation.core.validation.business
-
-import de.rki.coronawarnapp.covidcertificate.validation.core.validation.EvaluatedDccRule
-
-interface BusinessValidation {
-    val acceptanceRules: Set<EvaluatedDccRule>
-    val invalidationRules: Set<EvaluatedDccRule>
-}
